Unset the variants attribute from the product Model. Closes #229
Remove variants from the product object, as variants will be called over the __get method.
This will be solved differently in butik version 4.

Why did the error occur?
Variants will be called over a magic method. As the variant Fieldtype does set a variants attribute (which always is null), the `__get` method will not get called and the variants will not be fetched correctly.